% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownFunctions.R
\name{left_join_warn}
\alias{left_join_warn}
\title{Left Join (with warning)}
\usage{
left_join_warn(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)
}
\arguments{
\item{by}{See documentation for \code{dplyr::left_join}}

\item{copy}{See documentation for \code{dplyr::left_join}}

\item{suffix}{See documentation for \code{dplyr::left_join}}

\item{...}{See documentation for \code{dplyr::left_join}}

\item{keep}{See documentation for \code{dplyr::left_join}}

\item{`x, }{y\verb{A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See documentation for}dplyr::left_join`}
}
\description{
Perform a dplyr::left_join, but warn if there's a discrepancy in the number of rows between the argument x and the final joined df.
}
