% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownFunctions.R
\name{locate.nas}
\alias{locate.nas}
\title{Find and tally NA's in a data frame.}
\usage{
locate.nas(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with columns \code{colname} (column names from the original \code{df}) and \code{nas} (the number of NA's found in that column.)
}
\description{
Find and tally NA's in a data frame.
}
