% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webFunctions.R
\name{named_group_split}
\alias{named_group_split}
\title{Named Group Split}
\usage{
named_group_split(.tbl, ...)
}
\arguments{
\item{.tbl}{A data frame.}

\item{...}{Variables to group by}
}
\value{
A list, such that the names of the list are the group keys of the grouping variable(s).
}
\description{
This is a named version of dplyr::group_split, where the resulting list is named according to the variable it was split on. The function was written by Romain Francois, and I found it here: https://github.com/tidyverse/dplyr/issues/4223
}
