% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownFunctions.R
\name{nospace}
\alias{nospace}
\title{Remove all spaces from strings}
\usage{
nospace(vec)
}
\arguments{
\item{vec}{A vector. Will be coerced to character if it isn't already.}
}
\value{
A vector with no spaces.
}
\description{
Remove all spaces from strings
}
