% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownFunctions.R
\name{splitup}
\alias{splitup}
\title{Split Up}
\usage{
splitup(df, k)
}
\arguments{
\item{df}{A data frame.}

\item{k}{The number of groups/sets to split the data frame into.}
}
\value{
The data frame, with an added column \code{grp} containing group assignments.
}
\description{
Split a data frame for cross-validation
}
